Add MP-JWT supported claims to openid scope
MP-JWT 1.0 specification has introduced two claims upn and groups which are mandatory to generate a JWT token which is supported by MicroProfile JWT authentication framework.

This commit adds the two claims upn and grops under openid scope so that a user will be able to generate a JWT token in MP-JWT supported format.